migrating stats of clients and artists
