+++ src/app/profile/artists/stats/service/statsArtistService.ts
@@ -1,0 +1,21 @@
+import Cookies from 'js-cookie';
+
+export default async function statsArtistService() {
+	const token = Cookies.get('sillusr');
+	let userData;
+	try {
+		const res = await fetch(process.env.NEXT_PUBLIC_GET_UPDATE_DATAUSER as string, {
+			method: 'GET',
+			headers: {
+				Authorization: `Bearer ${token}`,
+			},
+			cache: 'no-store',
+		});
+		userData = await res.json();
+		return {
+			user: userData.data,
+		};
+	} catch (e) {
+		userData = { data: 'Token has expired' };
+	}
+}
