+++ src/styles/ArtistStats.module.scss
@@ -1,6 +1,12 @@
 .statsContainer {
-  margin: 0 auto;
-  margin-top: 75px;
-  padding: 2rem 3rem;
-  max-width: 1200px;
+	margin: 0 auto;
+	margin-top: 75px;
+	padding: 2rem 3rem;
+	max-width: 1200px;
+}
+
+.loader {
+	display: flex;
+	justify-content: center;
+	align-items: center;
 }
