+++ src/styles/Client.module.scss
@@ -13,3 +13,12 @@
 	justify-content: center;
 	align-items: center;
 }
+
+.posters {
+	width: 100%;
+	padding: 2rem;
+
+	@media screen and (min-width: '800px') {
+		width: 70%;
+	}
+}
