+++ src/styles/Layout.module.scss
@@ -1,0 +1,3 @@
+.main {
+	flex-grow: 1;
+}
