+++ src/styles/StatsPage.module.scss
@@ -1,3 +1,9 @@
 .statsContainer {
-  padding: 0.5rem 1rem;
+	padding: 0.5rem 1rem;
+}
+
+.loader {
+	display: flex;
+	justify-content: center;
+	align-items: center;
 }
